Feature: Add support to generate multiple types of certificates
