+++ .gitignore
@@ -72,6 +72,7 @@ typings/
 .env
 .env.test
 config.json
+config/
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
